refactor crud, and move code around
